Finished reading _The Deadline_
This was a cute and engaging book about managing software projects. 
Mr. Tompkins is laid off (or "set free to find more opportunities") and
is kidnapped to an ex-Soviet country, where he is given the opportunity
of a lifetime to lead a set of software projects and experiment with
different project environments. The book is chock full of important
advice, which Mr. Tompkins conveniently writes into his journal in
bullet-point form at the end of every chapter. Some good example:

* Pressure does not affect how fast a project proceeds. Pressure does
not make anyone think faster.
* Start every meeting by excusing some people whose time is extremely
valuable, after hearing a report or something. It should be made known
that meetings are meant to be lean and relevant.
* Conflicts should be respected, since they represent competing
project goals as seen by different people trying to accomplish good.
Having a process for mediation with trained mediators will help every
team in the organization get things done faster.
* Design should be done by a small team to keep it simple and the work
partitioning both practical and relevant to the final design.

Overall, charming.