finished The Systems Bible
Not what I expected, but still very relevant.
I expected something very academic and
mathematical, but instead it claimed that
but was far from it. The interesting thing,
though, is that it all still rings perfectly true.
A system can be a blessing or a curse, but
it is guaranteed to have unexpected behavior. 
When it does something bad, you'd better 
hope that your system is flexible, changable,
monitorable objectively somehow, that it doesn't
completely dominate everything and only
allow positive feedback.

Kind of a mix of Taoist philosophy, design, a tiny
bit of math (really, barely any), common sense,
self-improvement, endearing Latin textbook, etc.
Lately I'd been thinking about all of the generalizable
things I'd learned from programmign: that things need
to be flexible and interchangable, testable and constantly
tested, designed with user experience in mind, tested before
deployed, etc. All of that is abstracted out of software design
and into the real world with this book, which is really quite
phenomenal. Software happens to be a type of system.